+++ packages/subscription-controller/src/SubscriptionController.ts
@@ -8,6 +8,7 @@ import { StaticIntervalPollingController } from '@metamask/polling-controller';
 import type { AuthenticationController } from '@metamask/profile-sync-controller';
 
 import {
+  ACTIVE_SUBSCRIPTION_STATUSES,
   controllerName,
   DEFAULT_POLLING_INTERVAL,
   SubscriptionControllerErrorMessage,
@@ -23,7 +24,9 @@ import type {
   TokenPaymentInfo,
   UpdatePaymentMethodCardResponse,
   UpdatePaymentMethodOpts,
-  CachedLastSelectedPaymentMethods,
+  CachedLastSelectedPaymentMethod,
+  SubmitSponsorshipIntentsMethodParams,
+  RecurringInterval,
 } from './types';
 import {
   PAYMENT_TYPES,
@@ -47,7 +50,7 @@ export type SubscriptionControllerState = {
    */
   lastSelectedPaymentMethod?: Record<
     ProductType,
-    CachedLastSelectedPaymentMethods
+    CachedLastSelectedPaymentMethod
   >;
 };
 
@@ -89,6 +92,11 @@ export type SubscriptionControllerGetBillingPortalUrlAction = {
   handler: SubscriptionController['getBillingPortalUrl'];
 };
 
+export type SubscriptionControllerSubmitSponsorshipIntentsAction = {
+  type: `${typeof controllerName}:submitSponsorshipIntents`;
+  handler: SubscriptionController['submitSponsorshipIntents'];
+};
+
 export type SubscriptionControllerGetStateAction = ControllerGetStateAction<
   typeof controllerName,
   SubscriptionControllerState
@@ -103,7 +111,8 @@ export type SubscriptionControllerActions =
   | SubscriptionControllerGetCryptoApproveTransactionParamsAction
   | SubscriptionControllerStartSubscriptionWithCryptoAction
   | SubscriptionControllerUpdatePaymentMethodAction
-  | SubscriptionControllerGetBillingPortalUrlAction;
+  | SubscriptionControllerGetBillingPortalUrlAction
+  | SubscriptionControllerSubmitSponsorshipIntentsAction;
 
 export type AllowedActions =
   | AuthenticationController.AuthenticationControllerGetBearerToken
@@ -291,6 +300,11 @@ export class SubscriptionController extends StaticIntervalPollingController()<
     this.messenger.registerActionHandler(
       'SubscriptionController:getBillingPortalUrl',
       this.getBillingPortalUrl.bind(this),
+    );
+
+    this.messenger.registerActionHandler(
+      `${controllerName}:submitSponsorshipIntents`,
+      this.submitSponsorshipIntents.bind(this),
     );
   }
 
@@ -515,14 +529,14 @@ export class SubscriptionController extends StaticIntervalPollingController()<
    */
   cacheLastSelectedPaymentMethod(
     product: ProductType,
-    paymentMethod: CachedLastSelectedPaymentMethods,
+    paymentMethod: CachedLastSelectedPaymentMethod,
   ) {
     if (
       paymentMethod.type === PAYMENT_TYPES.byCrypto &&
-      !paymentMethod.paymentTokenAddress
+      (!paymentMethod.paymentTokenAddress || !paymentMethod.paymentTokenSymbol)
     ) {
       throw new Error(
-        SubscriptionControllerErrorMessage.PaymentTokenAddressRequiredForCrypto,
+        SubscriptionControllerErrorMessage.PaymentTokenAddressAndSymbolRequiredForCrypto,
       );
     }
 
@@ -531,6 +545,60 @@ export class SubscriptionController extends StaticIntervalPollingController()<
         ...state.lastSelectedPaymentMethod,
         [product]: paymentMethod,
       };
+    });
+  }
+
+  /**
+   * Submit sponsorship intents to the Subscription Service backend.
+   *
+   * This is intended to be used together with the crypto subscription flow.
+   * When the user has enabled the smart transaction feature, we will sponsor the gas fees for the subscription approval transaction.
+   *
+   * @param request - Request object containing the address and products.
+   * @example {
+   *   address: '0x1234567890123456789012345678901234567890',
+   *   products: [ProductType.Shield],
+   *   recurringInterval: RecurringInterval.Month,
+   *   billingCycles: 1,
+   * }
+   */
+  async submitSponsorshipIntents(
+    request: SubmitSponsorshipIntentsMethodParams,
+  ) {
+    if (request.products.length === 0) {
+      throw new Error(
+        SubscriptionControllerErrorMessage.SubscriptionProductsEmpty,
+      );
+    }
+
+    this.#assertIsUserNotSubscribed({ products: request.products });
+
+    // verify if the user has trailed the provided products before
+    const hasTrailedBefore = this.state.trialedProducts.some((product) =>
+      request.products.includes(product),
+    );
+    // if the user has not trialed the provided products before, submit the sponsorship intents
+    if (hasTrailedBefore) {
+      return;
+    }
+
+    const selectedPaymentMethod =
+      this.state.lastSelectedPaymentMethod?.[request.products[0]];
+    this.#assertIsPaymentMethodCrypto(selectedPaymentMethod);
+
+    const { paymentTokenSymbol, plan } = selectedPaymentMethod;
+    const productPrice = this.#getProductPriceByProductAndPlan(
+      // we only support one product at a time for now
+      request.products[0],
+      plan,
+    );
+    const billingCycles = productPrice.minBillingCycles;
+
+    await this.#subscriptionService.submitSponsorshipIntents({
+      ...request,
+      paymentTokenSymbol,
+      billingCycles,
+      recurringInterval: plan,
     });
   }
 
@@ -605,16 +673,6 @@ export class SubscriptionController extends StaticIntervalPollingController()<
     return tokenAmount.toString();
   }
 
-  #assertIsUserNotSubscribed({ products }: { products: ProductType[] }) {
-    if (
-      this.state.subscriptions.find((subscription) =>
-        subscription.products.some((p) => products.includes(p.name)),
-      )
-    ) {
-      throw new Error(SubscriptionControllerErrorMessage.UserAlreadySubscribed);
-    }
-  }
-
   /**
    * Triggers an access token refresh.
    */
@@ -625,6 +683,34 @@ export class SubscriptionController extends StaticIntervalPollingController()<
     this.messenger.call('AuthenticationController:performSignOut');
   }
 
+  #getProductPriceByProductAndPlan(
+    product: ProductType,
+    plan: RecurringInterval,
+  ): ProductPrice {
+    const { pricing } = this.state;
+    const productPricing = pricing?.products.find((p) => p.name === product);
+    const productPrice = productPricing?.prices.find(
+      (p) => p.interval === plan,
+    );
+    if (!productPrice) {
+      throw new Error(SubscriptionControllerErrorMessage.ProductPriceNotFound);
+    }
+    return productPrice;
+  }
+
+  #assertIsUserNotSubscribed({ products }: { products: ProductType[] }) {
+    const subscription = this.state.subscriptions.find((sub) =>
+      sub.products.some((p) => products.includes(p.name)),
+    );
+
+    if (
+      subscription &&
+      ACTIVE_SUBSCRIPTION_STATUSES.includes(subscription.status)
+    ) {
+      throw new Error(SubscriptionControllerErrorMessage.UserAlreadySubscribed);
+    }
+  }
+
   #assertIsUserSubscribed(request: { subscriptionId: string }) {
     if (
       !this.state.subscriptions.find(
@@ -632,6 +718,27 @@ export class SubscriptionController extends StaticIntervalPollingController()<
       )
     ) {
       throw new Error(SubscriptionControllerErrorMessage.UserNotSubscribed);
+    }
+  }
+
+  /**
+   * Asserts that the value is a valid crypto payment method.
+   *
+   * @param value - The value to assert.
+   * @throws an error if the value is not a valid crypto payment method.
+   */
+  #assertIsPaymentMethodCrypto(
+    value: CachedLastSelectedPaymentMethod | undefined,
+  ): asserts value is Required<CachedLastSelectedPaymentMethod> {
+    if (
+      !value ||
+      value.type !== PAYMENT_TYPES.byCrypto ||
+      !value.paymentTokenAddress ||
+      !value.paymentTokenSymbol
+    ) {
+      throw new Error(
+        SubscriptionControllerErrorMessage.PaymentMethodNotCrypto,
+      );
     }
   }